+++ src/components/study-group-detail/modal/index.ts
@@ -3,3 +3,4 @@ export { default as EditScheduleModal } from '@components/study-group-detail/mod
 export { default as ScheduleDatePickerModal } from '@components/study-group-detail/modal/ScheduleDatePickerModal'
 export { default as ScheduleDetailModal } from '@components/study-group-detail/modal/ScheduleDetailModal'
 export { default as ScheduleDetailDiv } from '@components/study-group-detail/modal/ScheduleDetailDiv'
+export { default as ScheduleForm } from '@components/study-group-detail/modal/ScheduleForm'
